fix routes, break auth into separate route
