+++ src/routes/index.js
@@ -1,8 +1,11 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
+const authRoutes = require('./auth');
+
 
 router.use('/api', apiRoutes);
+router.use('/auth', authRoutes);
 
 router.use('/', (req, res) => {
     res.render('index', {
