+++ src/server.js
@@ -24,12 +24,12 @@ sequelize.sync({ force: false }).then(() => {
   app.engine('handlebars', _engine);
   app.set('view engine', 'handlebars');
   app.set('views', './src/views');
-
+  app.use(routes);
   app.use(express.static('src/dist'))
   app.use(session(sess));
   app.use(express.json());
   app.use(express.urlencoded({ extended: true }));
-  app.use(routes);
+  
   
 
   app.listen(PORT, () => console.log('Now listening'));
